refactor: 메인 페이지에서 패널 목록을 `InfinitePanelGrid`로 분리한다 (#196)
